message dispatcher moved into integration events folder
