+++ WhatBackend.EmailRenderService/IntegrationEvents/MessageDispatcher.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using EasyNetQ.AutoSubscribe;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace WhatBackend.EmailRenderService
+namespace WhatBackend.EmailRenderService.IntegrationEvents
 {
     public class MessageDispatcher : IAutoSubscriberMessageDispatcher
     {
